cliend send (annual interest rate, number of years, and loan amount)  and get Total payment and monthly payment
Write a server for a client. The client sends loan information (annual interest rate, number of years, and loan amount) to the server  see the following figure. The server computes monthly payment and total payment and sends them back to the client. Name the client lab_1Client and the server lab_1Server